Remove old compoennts related to return + refund functionality
