+++ src/orders/components/OrderGrantRefundPage/messages.ts
@@ -1,4 +1,5 @@
-import { defineMessages } from "react-intl";
+import { FulfillmentStatus } from "@saleor/graphql";
+import { defineMessages, IntlShape } from "react-intl";
 
 export const grantRefundPageMessages = defineMessages({
   pageHeader: {
@@ -95,3 +96,37 @@ export const productCardMessages = defineMessages({
     description: "grant refund table, column header",
   },
 });
+
+export const fulfilmentStatusMessages = defineMessages({
+  fulfillment: {
+    id: "MewrtN",
+    defaultMessage: "Fulfillment",
+    description: "section header",
+  },
+  fulfillmentReturned: {
+    id: "H/f9KR",
+    defaultMessage: "Fulfillment returned",
+    description: "section header returned",
+  },
+  fulfillmentWaitingForApproval: {
+    id: "i/ZhxL",
+    defaultMessage: "Fulfillment waiting for approval",
+    description: "section header returned",
+  },
+});
+
+export const getTitle = (
+  fulfillmentStatus: FulfillmentStatus,
+  intl: IntlShape,
+) => {
+  switch (fulfillmentStatus) {
+    case FulfillmentStatus.RETURNED:
+      return intl.formatMessage(fulfilmentStatusMessages.fulfillmentReturned);
+    case FulfillmentStatus.WAITING_FOR_APPROVAL:
+      return intl.formatMessage(
+        fulfilmentStatusMessages.fulfillmentWaitingForApproval,
+      );
+    default:
+      return intl.formatMessage(fulfilmentStatusMessages.fulfillment);
+  }
+};
