--- src/orders/components/OrderRefund/index.ts
@@ -1,2 +1,0 @@
-export * from "./OrderRefund";
-export { default } from "./OrderRefund";
