--- src/orders/components/OrderRefundFulfilledProducts/index.ts
@@ -1,2 +1,0 @@
-export * from "./OrderRefundFulfilledProducts";
-export { default } from "./OrderRefundFulfilledProducts";
