--- src/orders/components/OrderRefundPage/index.ts
@@ -1,2 +1,0 @@
-export * from "./OrderRefundPage";
-export { default } from "./OrderRefundPage";
