--- src/orders/components/OrderRefundReturnAmount/index.ts
@@ -1,2 +1,0 @@
-export * from "./OrderRefundReturnAmount";
-export { default } from "./OrderRefundReturnAmount";
