--- src/orders/components/OrderRefundUnfulfilledProducts/index.ts
@@ -1,2 +1,0 @@
-export * from "./OrderRefundUnfulfilledProducts";
-export { default } from "./OrderRefundUnfulfilledProducts";
